# document/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Pokedex Application
  description: >-
    初代ポケモン151匹を表示するポケモン図鑑もどき。
paths:
...
  /task/tasks/{id}:
    get:
      summary: ミッション詳細表示
      description: |
      tags:
        - task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 【idが存在する】ミッション詳細表示に成功する。【それ以外】ミッション詳細表示に失敗する。/failsafe/failsafeのエラー画面に遷移する。
  /task/tasks/{id}/edit:
    get:
      summary: ミッション編集画面表示
      description: |
      tags:
        - task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 【idが存在する】ミッション編集画面表示に成功する。【それ以外】ミッション編集画面表示に失敗する。/failsafe/failsafeのエラー画面に遷移する。
  /task/tasks/update:
    post:
      summary: ミッション編集処理
      description: ミッション内容のみ更新可能。更新値に制限はなくは空欄でも更新する。
      tags:
        - task
      parameters:
        - in: query
          name: id
          required: false
          schema:
            type: integer
        - in: query
          name: detail
          required: false
          schema:
            type: string
      responses:
        '302':
          description: ミッション詳細の編集に成功する。task/tasks/{id}に遷移する。
  /task/tasks/delete/{id}:
    get:
      summary: ミッション削除処理
      description: |
      tags:
        - task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '302':
          description: 【idが存在する時】ミッション削除に成功する。/task/tasksに遷移する。【それ以外】ミッション削除に失敗する。/failsafe/failsafeのエラー画面に遷移する。
...
  /task/add:
    post:
      summary: ミッション登録処理
      description: ミッション内容のみ登録可能。更新値に制限はなくは空欄でも更新する。
      tags:
        - task
      parameters:
        - in: query
          name: addText
          required: true
          schema:
            type: string
      responses:
        '302':
          description: ミッション登録に成功する。task/tasksに遷移する。